[ALPS04461488] tmem: fix GCC build error
Fix GCC build error in kasan enabled project.

MTK-Commit-Id: 8b22c123dd8e7d885cf47e5db0eeb57bc064dc31

CR-Id: ALPS04461488
Feature: GenieZone

+++ drivers/misc/mediatek/trusted_mem/tee_devices.c
@@ -198,7 +198,7 @@ static int __init tee_smem_devs_init(void)
 	int idx = 0;
 
 	pr_info("%s:%d (%d)\n", __func__, __LINE__,
-		TEE_SECURE_MEM_DEVICE_COUNT);
+		(int)TEE_SECURE_MEM_DEVICE_COUNT);
 
 	for (idx = 0; idx < TEE_SECURE_MEM_DEVICE_COUNT; idx++) {
 		t_device = create_tee_smem_device(
